Refactor SDCARD; Separate generic code from SPI-specific code
